rename write to stringify and add stringifySync
