+++ src/index.ts
@@ -7,5 +7,5 @@ export * from './parseTimestamp'
 export * from './parseTimestamps'
 export * from './resync'
 export * from './stringify'
+export * from './stringifySync'
 export * from './types'
-export * from './write'
